Defining a handler interface, rearranging files
* Moving the local package to handlers/memcached/chunked
* Changing imports to regular go style vs. local package imports
